Nuevas correcciones y mejoras de varios puntos del tp1

+++ TP01_AmadorCristianEmanuel/src/ar/edu/unju/fi/ejercicio11/Main.java
@@ -4,9 +4,9 @@ public class Main {
 
 	public static void main(String[] args) {
 		
-		int j = 40;
+		byte j = 40;
 		
-		for (int i = 2; j > 0; i++) {
+		for (byte i = 2; j > 0; i++) {
 			
 			System.out.print(i*j + " ");
 			j--;
